app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
asyncpg==0.29.0
httpx==0.27.2
tenacity==9.0.0
prometheus-client==0.20.0
pyyaml==6.0.1
